refactor: remove reliance on griptape's SingletonMixin
In general we should avoid using the more "internal" aspects of
griptape.
