+++ src/griptape_nodes/exe_types/type_validator.py
@@ -3,14 +3,14 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from griptape.mixins.singleton_mixin import SingletonMixin
+from griptape_nodes.utils.metaclasses import SingletonMeta
 
 logger = logging.getLogger(__name__)
 
 ALLOWED_NUM_ARGS = 2
 
 
-class TypeValidator(SingletonMixin):
+class TypeValidator(metaclass=SingletonMeta):
     """A type string validator that checks against known types.
 
     Implemented as a singleton to ensure consistent behavior across an application.
