+++ src/griptape_nodes/utils/metaclasses.py
@@ -1,0 +1,10 @@
+from typing import Any, ClassVar
+
+
+class SingletonMeta(type):
+    _instances: ClassVar[dict] = {}
+
+    def __call__(cls, *args, **kwargs) -> Any:
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+        return cls._instances[cls]
